Refactored shell
