+++ budgeting-frontend/package.json
@@ -44,6 +44,7 @@
   },
   "devDependencies": {
     "@graphql-codegen/cli": "^2.16.3",
-    "@graphql-codegen/client-preset": "^1.2.5"
+    "@graphql-codegen/client-preset": "^1.2.5",
+    "prettier": "^2.8.2"
   }
 }
