+++ budgeting-frontend/styles/shell.style.ts
@@ -1,0 +1,33 @@
+import { createStyles } from "@mantine/core";
+
+const shellStyle = createStyles((theme, _params, getRef) => {
+  return {
+    link: {
+      ...theme.fn.focusStyles(),
+      display: "flex",
+      alignItems: "center",
+      textDecoration: "none",
+      fontSize: theme.fontSizes.sm,
+      color: theme.colorScheme === "dark" ? theme.colors.dark[1] : theme.colors.gray[7],
+      padding: `${theme.spacing.xs}px ${theme.spacing.xs}px`,
+      borderRadius: theme.radius.sm,
+      fontWeight: 500,
+
+      "&:hover": {
+        backgroundColor: theme.colorScheme === "dark" ? theme.colors.dark[6] : theme.colors.gray[0],
+        color: theme.colorScheme === "dark" ? theme.white : theme.black,
+      },
+    },
+    linkActive: {
+      "&, &:hover": {
+        backgroundColor: theme.fn.variant({ variant: "light", color: theme.primaryColor }).background,
+        color: theme.fn.variant({ variant: "light", color: theme.primaryColor }).color,
+      },
+    },
+    chevron: {
+      transition: "transform 200ms ease",
+    },
+  };
+});
+
+export default shellStyle;
